chore: rename status_activity for get_activity
