+++ backend/src/adapters/input/http/status_activity.py
@@ -4,7 +4,7 @@ from uuid import UUID
 
 from fastapi import APIRouter, Depends, Header, HTTPException, Path, status
 from src.adapters.input.http.schemas import ActivityResponse, StatusResponse
-from src.application.ports.input.status_activity_port import StatusActivityPort
+from src.application.ports.input.get_activity_port import StatusActivityPort
 from src.config.dependencies import get_status_activity_service
 
 logger = logging.getLogger(__name__)
