+++ backend/src/config/dependencies.py
@@ -9,7 +9,7 @@ from src.application.usecase.close_activity_service import CloseActivityService
 from src.application.usecase.create_activity_service import CreateActivityService
 from src.application.usecase.evaluation_activity_service import EvaluationActivityService
 from src.application.usecase.join_activity_service import JoinActivityService
-from src.application.usecase.status_activity_service import StatusActivityService
+from src.application.usecase.get_activity_service import StatusActivityService
 from src.config.settings import initialize_settings
 
 logger = logging.getLogger(__name__)
