+++ backend/tests/unit/application/test_status_activity_service.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock
 from uuid import UUID
 
 import pytest
-from src.application.usecase.status_activity_service import StatusActivityService
+from src.application.usecase.get_activity_service import StatusActivityService
 from src.domain.entities.participant import Participant
 
 ACTIVITY_ID = UUID("8e6587b8-b158-4068-a254-76bd0d31f4f7")
